feat: add ClientBottom componenent
