+++ packages/apps/crm/src/components/organisms/index.js
@@ -12,3 +12,4 @@ export {default as OpportunityBody} from './OpportunityBody/OpportunityBody';
 export {default as OpportunityBottom} from './OpportunityBottom/OpportunityBottom';
 export {default as ClientHeader} from './ClientHeader/ClientHeader';
 export {default as ClientBody} from './ClientBody/ClientBody';
+export {default as ClientBottom} from './ClientBottom/ClientBottom';
