+++ packages/apps/crm/src/screens/client/ClientDetailsScreen.js
@@ -1,12 +1,12 @@
 import React, {useEffect} from 'react';
-import {ScrollView, StyleSheet, View} from 'react-native';
-import {Screen, HeaderContainer, CircleButton} from '@axelor/aos-mobile-ui';
+import {ScrollView} from 'react-native';
+import {Screen, HeaderContainer} from '@axelor/aos-mobile-ui';
 import {
   useSelector,
   HeaderOptionsMenu,
   useDispatch,
 } from '@axelor/aos-mobile-core';
-import {ClientBody, ClientHeader} from '../../components';
+import {ClientBody, ClientBottom, ClientHeader} from '../../components';
 import {searchContactById} from '../../features/contactSlice';
 import {fetchPartnerEventById} from '../../features/eventSlice';
 import {getClientbyId} from '../../features/clientSlice';
@@ -50,52 +50,9 @@ const ClientDetailsScreen = ({navigation, route}) => {
       <ScrollView>
         <ClientBody navigation={navigation} />
       </ScrollView>
-      <View style={styles.bottomContainer}>
-        <CircleButton
-          iconName="pen"
-          onPress={() =>
-            navigation.navigate('ClientFormScreen', {
-              idClient: idClient,
-            })
-          }
-        />
-      </View>
+      <ClientBottom idClient={idClient} navigation={navigation} />
     </Screen>
   );
 };
-
-const styles = StyleSheet.create({
-  item: {
-    marginHorizontal: 12,
-    marginVertical: 4,
-  },
-  headerContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  headerContainerChildren: {
-    flexDirection: 'row',
-    marginLeft: '5%',
-    alignItems: 'center',
-  },
-  headerInfo: {
-    flexDirection: 'column',
-    marginLeft: '7%',
-  },
-  container: {
-    alignItems: 'center',
-  },
-  textTitle: {
-    fontWeight: 'bold',
-  },
-  bottomContainer: {
-    width: '90%',
-    flexDirection: 'row-reverse',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginVertical: 2,
-    marginBottom: 25,
-  },
-});
 
 export default ClientDetailsScreen;
